server/cv-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
easyocr
ultralytics